pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "microgpt"
version = "2.0.0"
description = "A comprehensive, pure-Python GPT ecosystem with state-of-the-art features"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "microgpt Team", email = "team@microgpt.ai"}
]
maintainers = [
    {name = "microgpt Team", email = "team@microgpt.ai"}
]
keywords = [
    "gpt", "transformer", "deep-learning", "nlp", "language-model",
    "pytorch-free", "pure-python", "llm", "ai", "machine-learning"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
web = [
    "flask>=2.0.0",
    "gunicorn>=20.0.0",
]
viz = [
    "matplotlib>=3.5.0",
    "plotly>=5.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]
export = [
    "torch>=1.10.0",
    "onnx>=1.12.0",
    "transformers>=4.20.0",
]
all = [
    "flask>=2.0.0",
    "gunicorn>=20.0.0",
    "matplotlib>=3.5.0",
    "plotly>=5.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "torch>=1.10.0",
    "onnx>=1.12.0",
    "transformers>=4.20.0",
]

[project.urls]
Homepage = "https://github.com/iamGodofall/karpathy-microgpt-by-Enock"
Documentation = "https://github.com/iamGodofall/karpathy-microgpt-by-Enock/blob/master/docs/GUIDE.md"
Repository = "https://github.com/iamGodofall/karpathy-microgpt-by-Enock"
Issues = "https://github.com/iamGodofall/karpathy-microgpt-by-Enock/issues"
Changelog = "https://github.com/iamGodofall/karpathy-microgpt-by-Enock/blob/master/CHANGELOG.md"

[project.scripts]
microgpt = "main:main"
microgpt-train = "microgpt.cli:train_cli"
microgpt-generate = "microgpt.cli:generate_cli"
microgpt-chat = "microgpt.chat:main"
microgpt-server = "microgpt.api_server:main"


[tool.setuptools]
packages = ["microgpt"]

[tool.setuptools.package-data]
microgpt = ["*.yaml", "*.json", "*.md"]

[tool.setuptools.package-dir]
microgpt = "microgpt"


[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
    "integration_test.py",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "examples/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
